requirements.txt
slack-bolt>=1.18.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
openai>=1.3.0
sqlalchemy>=2.0.23
python-decouple>=3.8